% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsNumeric}
\alias{AsNumeric}
\title{AsNumeric}
\usage{
AsNumeric(x, return_NA = FALSE)
}
\arguments{
\item{x}{Vector.}

\item{return_NA}{Logical.
If \code{return_NA = TRUE},
returns \code{NA} when \code{x} cannot be coerced.
If \code{return_NA = FALSE},
returns the original item when \code{x} cannot be coerced.}
}
\description{
By default, the function converts the input to numeric
if it can be coerced.
Otherwise, the original input is returned.
If all elements in the vector can be coerced to numeric,
the resulting vector will be numeric.
}
\examples{
x <- c("1", "2", "3")
AsNumeric(x)
x <- c("1", "2", "3", "a")
AsNumeric(x, return_NA = FALSE)
AsNumeric(x, return_NA = TRUE)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{utility functions}
\keyword{utils}
