% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C}
\alias{C}
\alias{C.default}
\alias{C.yac_symbol}
\title{Matrix of Covariance Expectations \eqn{\mathbf{C}}}
\usage{
C(A, S, check = TRUE, ...)

\method{C}{default}(A, S, check = TRUE, ...)

\method{C}{yac_symbol}(
  A,
  S,
  check = TRUE,
  exe = TRUE,
  R = FALSE,
  format = "ysym",
  simplify = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t by t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{check}{Logical.
If \code{check = TRUE} do some preprocessing
with input matrices using \code{\link[=CheckRAMMatrices]{CheckRAMMatrices()}}.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE},
executes the resulting \code{yacas} expression.
If \code{exe = FALSE},
returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE},
the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
\eqn{
  \mathbf{C} = \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}}
}
}
\description{
Derives the matrix of covariance expectations \eqn{\mathbf{C}}.
}
\details{
The matrix of covariance expectations \eqn{\mathbf{C}}
as a function of Reticular Action Model (RAM) matrices is given by

\deqn{
    \mathbf{C}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}} \\
    =
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
C(A, S)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigmax2", "sigmae2")
C(Ryacas::ysym(A), S)
C(Ryacas::ysym(A), S, format = "str")
C(Ryacas::ysym(A), S, format = "tex")
C(Ryacas::ysym(A), S, R = TRUE)

# Assigning values to symbols

beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1

C(Ryacas::ysym(A), S)
C(Ryacas::ysym(A), S, format = "str")
C(Ryacas::ysym(A), S, format = "tex")
C(Ryacas::ysym(A), S, R = TRUE)
eval(C(Ryacas::ysym(A), S, R = TRUE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{RAMScaled}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
