% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C_num}
\alias{C_num}
\title{Matrix of Covariance Expectations
\eqn{\mathbf{C}} - Numeric}
\usage{
C_num(A, S)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix
\eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}
}
\description{
Derives the matrix of covariance expectations
\eqn{\mathbf{C}}.
}
\details{
The matrix of covariance expectations
\eqn{\mathbf{C}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{C}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}} \\
    =
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- 1
A[1, 3] <- 1
diag(S) <- c(0, 0.25, 0.25)
colnames(A) <- rownames(A) <- c("y", "x", "e")
C_num(A, S)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other C functions: 
\code{\link{C_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{C functions}
\keyword{C}
