% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C_sym}
\alias{C_sym}
\title{Matrix of Covariance Expectations
\eqn{\mathbf{C}} - Symbolic}
\usage{
C_sym(A, S, simplify = FALSE)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix
\eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{simplify}{Logical.
Simplify the results.}
}
\description{
Derives the matrix of covariance expectations
\eqn{\mathbf{C}}.
}
\details{
The matrix of covariance expectations
\eqn{\mathbf{C}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{C}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}} \\
    =
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- S <- matrixR::zeroes(3, 3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigma2x", "sigma2e")
C_sym(A, S)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other C functions: 
\code{\link{C_num}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{C functions}
\keyword{C}
