% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRAMMatrices.R
\name{CheckRAMMatrices}
\alias{CheckRAMMatrices}
\alias{CheckRAMMatrices.default}
\alias{CheckRAMMatrices.yac_symbol}
\title{Check Matrices}
\usage{
CheckRAMMatrices(
  A,
  S = NULL,
  u = NULL,
  Filter = NULL,
  C = NULL,
  C.scaled = NULL,
  v = NULL,
  ...
)

\method{CheckRAMMatrices}{default}(
  A,
  S = NULL,
  u = NULL,
  Filter = NULL,
  C = NULL,
  C.scaled = NULL,
  v = NULL,
  ...
)

\method{CheckRAMMatrices}{yac_symbol}(
  A,
  S = NULL,
  u = NULL,
  Filter = NULL,
  C = NULL,
  C.scaled = NULL,
  v = NULL,
  ...
)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t by t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{Filter}{\verb{p by t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{C}{\verb{t by t} numeric matrix \eqn{\mathbf{C}}.
Model-implied variance-covariance matrix.}

\item{C.scaled}{\verb{t by t} numeric matrix
\eqn{\mathbf{C}_{\mathrm{scaled}}}.
Scaled/standardized model-implied variance-covariance matrix.}

\item{v}{vector of length \code{t} or \verb{t by 1} matrix.
Expected values.}

\item{...}{...}
}
\description{
Checks if the input matrices are in the correct form.
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
diag(S) <- c(0, 0.25, 1)
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
C <- matrix(
  data = c(
    1.25, 0.25, 1.00,
    0.25, 0.25, 0.00,
    1.00, 0.00, 1.00
  ),
  nrow = dim(A)[1]
)
v <- c(0.50, 0.50, 0.00)
CheckRAMMatrices(
  A = A,
  S = S,
  u = u,
  Filter = Filter,
  C = C,
  v = v
)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigmax2", "sigmae2")
u <- c("alpha", "mux", 0)
C <- matrix(
  data = c(
    "sigmax2*beta^2+sigmae2", "sigmax2*beta", "sigmae2",
    "sigmax2*beta", "sigmax2", 0,
    "sigmae2", 0, "sigmae2"
  ),
  nrow = dim(A)[1]
)
v <- c("alpha+beta*mux", "mux", 0)
CheckRAMMatrices(
  A = Ryacas::ysym(A),
  S = S,
  u = u,
  Filter = Filter,
  C = C,
  v = v
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{preprocess functions}
\keyword{preprocess}
