% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E_sym}
\alias{E_sym}
\title{Matrix of Total Effects
\eqn{\mathbf{E}} - Symbolic}
\usage{
E_sym(A)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}
}
\value{
Returns the matrix of total effects
\eqn{\mathbf{E}}.
}
\description{
Derives the matrix of total effects
\eqn{\mathbf{E}}.
}
\details{
The matrix of total effects
\eqn{\mathbf{E}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{E}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- "beta"
A[1, 3] <- 1
E_sym(A)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other E functions: 
\code{\link{E_num}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{E functions}
\keyword{E}
