% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq2Data.R
\name{Eq2Data}
\alias{Eq2Data}
\title{Equations to Multivariate Normal Sample Data}
\usage{
Eq2Data(eq, n, ...)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{n}{Integer.
Sample size.}

\item{...}{Additional arguments to pass to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\value{
\eqn{n} variates from \eqn{ \mathbf{X} \sim \mathcal{N}_{k}
    \left(
      \boldsymbol{\mu} \left( \boldsymbol{\theta} \right),
      \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    \right) .
  }
}
\description{
Generates data from a multivariate normal distribution
from model equations using the \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}} function.
}
\details{
The input is a character string
that specifies the associations between the variables.

The multivariate normal distribution is given by
\deqn{
    \mathbf{X} \sim \mathcal{N}_{k}
    \left(
      \boldsymbol{\mu},
      \boldsymbol{\Sigma}
    \right)
  }

with location parameter

\deqn{
    \boldsymbol{\mu} \in \mathbf{R}^{k}
  }

and a positive definite covariance matrix

\deqn{
    \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} .
  }

The probability density function is given by

\deqn{
    f_{\mathbf{X}} \left( x_1, \cdots, x_k \right)
    =
    \frac{
      \exp
      \left[
        - \frac{1}{2}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)^{\mathsf{T}}
        \boldsymbol{\Sigma}^{-1}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)
      \right]
      }{
        \sqrt{
          \left( 2 \pi \right)^k
          | \boldsymbol{\Sigma} |
        }
      }
    }

In this function,
the model-implied mean vector and variance-covariance matrix
are used as parameters to generate the data.

\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u}
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u}
  }
\deqn{
    \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{C}
    \mathbf{F}^{\mathsf{T}}
  }
}
\section{Syntax}{

Each line should follow the syntax below

\verb{lhs <space> op <space> rhs <space> par.label <\\n> or <;>}

\describe{
\item{lhs}{
is the variable on the \strong{left-hand side},
}
\item{rhs}{
is the variable on the \strong{right-hand side},
}
\item{op}{
is the \strong{operation} between \code{lhs} and \code{rhs},
}
\item{par.label}{
is the column of \strong{parameter label},
}
\item{\\n or ;}{
are \strong{line breaks}.
\strong{Each line should end with a line break.}
}
}

The associations are defined by the following operations

\describe{
\item{by}{
\verb{left-hand side} measured \strong{by} \verb{right-hand side},
}
\item{on}{
\verb{left-hand side} regressed \strong{on} \verb{right-hand side},
}
\item{with}{
\verb{left-hand side} covarying \strong{with} \verb{right-hand side},
}
\item{on 1}{
\verb{left-hand side} regressed \strong{on 1} for mean structure.
}
}
}

\section{par.label}{

Each parameter should have a numeric value.
}

\section{Line breaks}{

The characters \verb{\\n} and \verb{;} can be used as line breaks.
\strong{Each line should end with a line break.}
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
eq <- "
  # lhs op   rhs value
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2Data(eq, n = 100)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other data generation functions: 
\code{\link{RAM2Data}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generation functions}
\keyword{data}
