% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq2exp.R
\name{eq2exp_num}
\alias{eq2exp_num}
\title{Equations to Expectations - Numeric}
\usage{
eq2exp_num(eq)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}
}
\description{
Converts equations to expectations.
}
\details{
The input is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{VARIABLE1 OPERATION VARIABLE2 VALUE}

The associations are defined by the following operations

\describe{
\item{by}{\code{VARIABLE1} measured \strong{by} \code{VARIABLE2}}
\item{on}{\code{VARIABLE1} regressed \strong{on} \code{VARIABLE2}}
\item{with}{\code{VARIABLE1} covarying \strong{with} \code{VARIABLE2}}
\item{on 1}{\code{VARIABLE1} regressed \strong{on 1} for mean structure}
}
}

\section{Value}{

Each parameter should have a numeric value.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
eq <- "
  # VARIABLE1 OPERATION VARIABLE2 VALUE
  e           by        y         1.00
  y           on        x         1.00
  e           with      e         0.25
  x           with      x         0.25
  y           on        1         0.00
  x           on        1         0.50
"
eq2exp_num(eq)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
