% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq2exp.R
\name{eq2exp_sym}
\alias{eq2exp_sym}
\title{Equations to Expectations - Symbolic}
\usage{
eq2exp_sym(eq, simplify = FALSE)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{simplify}{Logical.
Simplify the results.}
}
\description{
Converts equations to expectations.
}
\details{
The input is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{VARIABLE1 OPERATION VARIABLE2 LABEL/VALUE}

The associations are defined by the following operations

\describe{
\item{by}{\code{VARIABLE1} measured \strong{by} \code{VARIABLE2}}
\item{on}{\code{VARIABLE1} regressed \strong{on} \code{VARIABLE2}}
\item{with}{\code{VARIABLE1} covarying \strong{with} \code{VARIABLE2}}
\item{on 1}{\code{VARIABLE1} regressed \strong{on 1} for mean structure}
}
}

\section{Label}{

Each parameter should be labeled.
The \code{LABEL} should be a number for fixed parameters
and a variable name for free parameters.
Equality contraints can be imposed by using the same variable name.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
eq <- "
  # VARIABLE1 OPERATION VARIABLE2 LABEL/VALUE
  e           by        y         1
  y           on        x         beta
  e           with      e         varofe
  x           with      x         varofx
  y           on        1         alpha
  x           on        1         meanofx
"
eq2exp_sym(eq)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
