% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq2Expectations.R
\name{Eq2Expectations}
\alias{Eq2Expectations}
\title{Equations to Expectations}
\usage{
Eq2Expectations(
  eq,
  par = FALSE,
  check = TRUE,
  R = FALSE,
  format = "ysym",
  simplify = FALSE
)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{par}{Logical.
If \code{par = TRUE}, use \code{par.index} as labels.
If \code{par = FALSE}, use \code{par.label} as labels.}

\item{check}{Logical.
If \code{check = TRUE} do some preprocessing
with input matrices using \code{\link[=CheckRAMMatrices]{CheckRAMMatrices()}}.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
Returns a list with the following elements

\describe{
\item{par.table}{Parameter table.}
\item{variables}{Variable names.}
\item{g.variables}{Variable names of observed variables.}
\item{h.variables}{Variable names of latent variables.}
\item{A}{
\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.
}
\item{S}{
\verb{t by t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.
}
\item{u}{
\verb{t by 1} matrix \eqn{\mathbf{u}} of mean structure parameters.
}
\item{Filter}{
\verb{p by t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.
}
\item{v}{
\verb{t by 1} matrix \eqn{\mathbf{v}}
of expected values.
}
\item{g}{
\verb{p by 1} matrix \eqn{\mathbf{g}}
of expected values of observed variables.
}
\item{C}{
\verb{t by t} matrix \eqn{\mathbf{C}}
of expected covariances.
}
\item{M}{
\verb{p by p} matrix \eqn{\mathbf{M}}
of expected covariances of observed variables.
}
}
}
\description{
Converts equations to expectations.
The argument \code{eq} is a character string
that specifies the associations between the variables.
See \code{Syntax}, \code{Operations}, \code{par.label},
\verb{Line breaks}, and \code{Comments} below.
}
\details{
The vector of expected values \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{v}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{E}
    \mathbf{u}
  }

The vector of expected values
of observed variables \eqn{\mathbf{g}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{v}
  }

The matrix of covariance expectations \eqn{\mathbf{C}}
as a function of Reticular Action Model (RAM) matrices is given by

\deqn{
    \mathbf{C}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}} \\
    =
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
  }

The matrix of covariance expectations
for given variables \eqn{\mathbf{M}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{C}
    \mathbf{F}^{\mathsf{T}}
  }

The matrix of scaled/standardized covariance expectations
\eqn{\mathbf{C}_{\mathrm{scaled}}}
(also known as correlations)
is given by

\deqn{
    \mathbf{C}_{\mathrm{scaled}}
    =
    \mathbf{D}^{-1}
    \mathbf{C}
    \mathbf{D}^{-1}
  }

The matrix of scaled/standardized covariance expectations
for given variables
\eqn{\mathbf{M}_{\mathrm{scaled}}}
(also known as correlations)
is given by

\deqn{
    \mathbf{M}_{\mathrm{scaled}}
    =
    \mathbf{F}
    \mathbf{C}_{\mathrm{scaled}}
    \mathbf{F}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters
for the mean structure,
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{D}_{t \times t}} represents a diagonal matrix
who diagonal elements are the square root of the diagonal elements
of \eqn{\mathbf{C}},
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables,
that is \eqn{p + q} .
}
}
\section{Syntax}{

Each line should follow the syntax below

\verb{lhs <space> op <space> rhs <space> par.label <\\n> or <;>}

\describe{
\item{lhs}{
is the variable on the \strong{left-hand side},
}
\item{rhs}{
is the variable on the \strong{right-hand side},
}
\item{op}{
is the \strong{operation} between \code{lhs} and \code{rhs},
}
\item{par.label}{
is the column of \strong{parameter label},
}
\item{\\n or ;}{
are \strong{line breaks}.
\strong{Each line should end with a line break.}
}
}
}

\section{Operations}{

The associations are defined by the following operations

\describe{
\item{by}{
\verb{left-hand side} measured \strong{by} \verb{right-hand side},
}
\item{on}{
\verb{left-hand side} regressed \strong{on} \verb{right-hand side},
}
\item{with}{
\verb{left-hand side} covarying \strong{with} \verb{right-hand side},
}
\item{on 1}{
\verb{left-hand side} regressed \strong{on 1} for mean structure.
}
}
}

\section{par.label}{

Each parameter should be labeled.
The \code{par.label} should be a number for fixed parameters
and a character string for free parameters.
Equality contraints can be imposed by using the same \code{par.label}.
}

\section{Line breaks}{

The characters \verb{\\n} and \verb{;} can be used as line breaks.
\strong{Each line should end with a line break.}
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
# Numerical ---------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2Expectations(eq)

# Symbolic ----------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2Expectations(eq, par = FALSE)
Eq2Expectations(eq, par = TRUE)

# Expressions using `par.label`

beta <- 1
sigmae2 <- 1
sigmax2 <- 0.25
alpha <- 0
mux <- 0.50
Exp <- Eq2Expectations(eq, par = FALSE, R = TRUE)
eval(Exp$M)
eval(Exp$g)

# Expressions using `par.index`

p <- c(beta, sigmae2, sigmax2, alpha, mux)
p1 <- p[1]
p2 <- p[2]
p3 <- p[3]
p4 <- p[4]
p5 <- p[5]
Exp <- Eq2Expectations(eq, par = TRUE, R = TRUE)
eval(Exp$M)
eval(Exp$g)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other eq functions: 
\code{\link{Eq2RAM}()},
\code{\link{EqParse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{eq functions}
\keyword{eq}
