% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq2ram.R
\name{eq2ram}
\alias{eq2ram}
\title{Equations to RAM Matrices}
\usage{
eq2ram(model)
}
\arguments{
\item{model}{Character string. Input model. See Details.}
}
\description{
Converts model equations to RAM matrices.
}
\details{
The input model is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{VARIABLE1 OPERATION VARIABLE2 LABEL}

The associations are defined by the following operations

\describe{
\item{by}{\code{VARIABLE1} measured \strong{by} \code{VARIABLE2}}
\item{on}{\code{VARIABLE1} regressed \strong{on} \code{VARIABLE2}}
\item{with}{\code{VARIABLE1} covarying \strong{with} \code{VARIABLE2}}
\item{on 1}{\code{VARIABLE1} regressed \strong{on 1} for mean structure}
}

\strong{Each line should end with a semicolon (\verb{;}).}
}

\section{Label}{

Each parameter should be labeled.
The \code{LABEL} should be a number for fixed parameters
and a variable name for free parameters.
Equality contraints can be imposed by using the same variable name.
Variable names can be written
following the \code{R} mathematical notation \code{grDevices::plotmath()}.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
model <- "
  # VARIABLE1 OPERATION VARIABLE2 LABEL
  e           by        y         1;
  y           on        x         beta;
  e           with      e         sigma[varepsilon]^2;
  x           with      x         sigma[x]^2;
  y           on        1         alpha;
  x           on        1         mu[x]
"
eq2ram(model)
}
\author{
Ivan Jacob Agaloos Pesigan
}
