% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expectations.R
\name{Expectations}
\alias{Expectations}
\alias{Expectations.default}
\alias{Expectations.yac_symbol}
\title{Expectations from the Reticular Action Model (RAM) Matrices}
\usage{
Expectations(A, S, u = NULL, Filter = NULL, ...)

\method{Expectations}{default}(
  A,
  S,
  u = NULL,
  Filter = NULL,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)

\method{Expectations}{yac_symbol}(
  A,
  S,
  u = NULL,
  Filter = NULL,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{Filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Derives the means and covariance expectations
from the Reticular Action Model (RAM) matrices.
}
\details{
The vector of mean structure parameters \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{u}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{v} \\
    =
    \mathbf{E}^{-1}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{v}_{t \times 1}} vector of expected values.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(Filter) <- c("y", "x", "e")
Expectations(A, S, u, Filter)

# Symbolic ----------------------------------------------------------
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigmax2", "sigmae2")
u <- c("alpha", "mux", 0)
Expectations(Ryacas::ysym(A), S, u, Filter)
Expectations(Ryacas::ysym(A), S, u, Filter, tex = TRUE)
Expectations(Ryacas::ysym(A), S, u, Filter, ysym = FALSE)
Expectations(Ryacas::ysym(A), S, u, Filter, str = FALSE)

alpha <- 0
beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1
mux <- 0.50
Expectations(Ryacas::ysym(A), S, u, Filter)
Expectations(Ryacas::ysym(A), S, u, Filter, tex = TRUE)
Expectations(Ryacas::ysym(A), S, u, Filter, ysym = FALSE)
Expectations(Ryacas::ysym(A), S, u, Filter, str = FALSE)
Expectations <- Expectations(Ryacas::ysym(A), S, u, Filter, str = FALSE)
eval(Expectations$A)
eval(Expectations$S)
eval(Expectations$u)
eval(Expectations$Filter)
eval(Expectations$v)
eval(Expectations$g)
eval(Expectations$C)
eval(Expectations$M)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
