% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\alias{g.default}
\alias{g.yac_symbol}
\title{Vector of Expected Values of Observed Variables \eqn{\mathbf{g}}}
\usage{
g(A, u, Filter = NULL, ...)

\method{g}{default}(A, u, Filter = NULL, ...)

\method{g}{yac_symbol}(
  A,
  u,
  Filter = NULL,
  exe = TRUE,
  R = FALSE,
  format = "ysym",
  simplify = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{Filter}{\verb{p by t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE},
executes the resulting \code{yacas} expression.
If \code{exe = FALSE},
returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE},
the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as laTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
\eqn{\mathbf{g} = \mathbf{F} \mathbf{v}}
}
\description{
Derives the vector of expected values
of observed variables \eqn{\mathbf{g}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values
of observed variables \eqn{\mathbf{g}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters
for the mean structure,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables,
that is \eqn{p + q} .
}
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(A) <- rownames(A) <- c("y", "x", "e")
g(A, u, Filter)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
u <- c("alpha", "mux", 0)
g(Ryacas::ysym(A), u, Filter)
g(Ryacas::ysym(A), u, Filter, format = "tex")
g(Ryacas::ysym(A), u, Filter, format = "ysym")
g(Ryacas::ysym(A), u, Filter, R = TRUE)

# Assigning values to symbols

alpha <- 0
beta <- 1
mux <- 0.50
g(Ryacas::ysym(A), u, Filter)
g(Ryacas::ysym(A), u, Filter, format = "tex")
g(Ryacas::ysym(A), u, Filter, format = "ysym")
g(Ryacas::ysym(A), u, Filter, R = TRUE)
eval(g(Ryacas::ysym(A), u, Filter, R = TRUE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
