% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g_num}
\alias{g_num}
\title{Vector of Expected Values of Observed Variables
\eqn{\mathbf{g}} - Numeric}
\usage{
g_num(A, u, filter)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}
}
\value{
Returns the vector of expected values of observed variables
\eqn{\mathbf{g}}.
}
\description{
Derives the vector of expected values of observed variables
\eqn{\mathbf{g}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values of observed variables
\eqn{\mathbf{g}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters for the mean structurem,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{p + q} .
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- 1
A[1, 3] <- 1
u <- c(0.00, 0.50, 0.00)
filter <- diag(2)
filter <- cbind(filter, 0)
colnames(filter) <- c("y", "x", "e")
rownames(filter) <- c("y", "x")
g_num(A, u, filter)
}
\seealso{
Other g functions: 
\code{\link{g_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{g functions}
\keyword{g}
