% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IminusA.R
\name{IminusA_num}
\alias{IminusA_num}
\title{\eqn{\mathbf{I} - \mathbf{A}} - Numeric}
\usage{
IminusA_num(A)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}
}
\value{
Returns \eqn{\mathbf{I} - \mathbf{A}}.
}
\description{
Derives \eqn{\mathbf{I} - \mathbf{A}}
}
\details{
\deqn{
    \mathbf{I} - \mathbf{A}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::zeroes(3, 3)
A[1, ] <- c(0, 1, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
IminusA_num(A)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other IminusA functions: 
\code{\link{IminusA_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{IminusA functions}
\keyword{IminusA}
