% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.R
\name{M_sym}
\alias{M_sym}
\title{Matrix of Covariance Expectations of Observed Variables
\eqn{\mathbf{M}} - Symbolic}
\usage{
M_sym(A, S, filter)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix
\eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}
}
\description{
Derives the matrix of covariance expectations of observed variables
\eqn{\mathbf{M}}.

\deqn{
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{C}
    \mathbf{F}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{p + q} .
}
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigma2x", "sigma2e")
filter <- diag(2)
filter <- cbind(filter, 0)
M_sym(A, S, filter)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other M functions: 
\code{\link{M_num}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{M functions}
\keyword{M}
