% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutheta.R
\name{mutheta}
\alias{mutheta}
\title{Model-Implied Mean Vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}}
\usage{
mutheta(m, A, filter)
}
\arguments{
\item{m}{\verb{t x 1} numeric vector \eqn{\mathbf{m}_{t \times 1}}.
Mean structure. Vector of means and intercepts.}

\item{A}{\verb{t x t} numeric matrix
\eqn{\mathbf{A}_{t \times t}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{filter}{\verb{j x t} numeric matrix
\eqn{\mathbf{F}_{j \times t}}.
Filter matrix used to select variables.}
}
\value{
Returns the model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
derived from the \code{M}, \code{A}, and \code{filter} matrices.
}
\description{
Derives the model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
using the Reticular Action Model (RAM) notation.
}
\details{
The model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{F}
    \left( \mathbf{I} - \mathbf{A} \right)^{-1}
    \mathbf{F}^{\mathsf{T}}
    \mathbf{m}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{F}_{j \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{m}_{t \times 1}} represents the mean structure,
that is, a vector of means and intercepts,
\item \eqn{j} number of observed variables,
\item \eqn{k} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{j + k} .
}
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules
for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald}
(pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{Omega}()},
\code{\link{Sigmatheta}()},
\code{\link{m}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
