% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Generate Data from a Multivariate Normal Distribution
Using the Reticular Action Model (RAM) Notation}
\usage{
mvn(n, A, Omega, M, filter = NULL, ...)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{A}{\verb{t x t} numeric matrix
\eqn{\mathbf{A}_{t \times t}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{Omega}{\verb{t x t} numeric matrix
\eqn{\boldsymbol{\Omega}_{t \times t}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{M}{\verb{t x 1} numeric vector \eqn{\mathbf{M}_{t \times 1}}.
Mean structure. Vector of means and intercepts.}

\item{filter}{\verb{j x t} numeric matrix
\eqn{\mathbf{F}_{j \times t}}.
Filter matrix used to select variables.
If \code{filter = NULL},
the filter matrix \eqn{\mathbf{F}} is assumed to be
a \eqn{t \times t} identity matrix.}

\item{...}{Additional arguments to pass to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\description{
Generates data from a multivariate normal distribution
using the Reticular Action Model (RAM) notation
with the \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}} function.
}
\details{
The Multivariate Normal Distribution is given by
\deqn{
    \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right)
  }

with location parameter

\deqn{
    \boldsymbol{\mu} \in \mathbb{R}^{k}
  }

and a positive definite covariance matrix

\deqn{
    \boldsymbol{\Sigma} \in \mathbb{R}^{k \times k} .
  }

The probability density function is given by

\deqn{
    f_{\mathbf{X}} \left( x_1, \cdots, x_k \right)
    =
    \frac{
      \exp
      \left[
        - \frac{1}{2}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)^{\mathsf{T}}
        \boldsymbol{\Sigma}^{-1}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)
      \right]
      }{
        \sqrt{
          \left( 2 \pi \right)^k
          | \boldsymbol{\Sigma} |
        }
      }
    }

In this function,
the model-implied mean vector and variance-covariance matrix
are used as parameters to generate the data.

\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{F}
    \left( \mathbf{I} - \mathbf{A} \right)^{-1}
    \mathbf{M}
  }
\deqn{
    \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{F}
    \left( \mathbf{I} - \mathbf{A} \right)^{-1}
    \boldsymbol{\Omega}
    \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}}
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
