% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega.R
\name{Omega}
\alias{Omega}
\title{Matrix of symmetric paths (double-headed arrows)
\eqn{\boldsymbol{\Omega}}}
\usage{
Omega(A, Sigmatheta)
}
\arguments{
\item{A}{\verb{t x t} numeric matrix
\eqn{\mathbf{A}_{t \times t}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{Sigmatheta}{\verb{t x t} numeric matrix
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}.
Model-implied variance-covariance matrix.}
}
\value{
Returns the matrix of symmetric paths (double-headed arrows)
\eqn{\boldsymbol{\Omega}}
derived from the \code{A} and \code{Sigmatheta} matrices.
}
\description{
Derives the matrix of symmetric paths (double-headed arrows)
\eqn{\boldsymbol{\Omega}}
using the Reticular Action Model (RAM) notation.
}
\details{
The matrix of symmetric paths (double-headed arrows)
\eqn{\boldsymbol{\Omega}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \boldsymbol{\Omega}
    =
    \left( \mathbf{I} - \mathbf{A} \right)
    \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\boldsymbol{\Sigma}_{t \times t}} represents
the model-implied variance-covariance matrix,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{j} number of observed variables,
\item \eqn{k} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{j + k} .
}
}
\references{
McArdle, J. J. (2005).
The development of the RAM rules
for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald}
(pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{E}()},
\code{\link{M}()},
\code{\link{Sigmatheta}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
