% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{Omega_linreg}
\alias{Omega_linreg}
\title{\eqn{\boldsymbol{\Omega}} Matrix from Linear Regression Model Parameters}
\usage{
Omega_linreg(sigma2, SigmaX)
}
\arguments{
\item{sigma2}{Numeric.
\eqn{\sigma^2} is the variance of the stochastic error term
\eqn{\boldsymbol{\varepsilon}}.}

\item{SigmaX}{Numeric matrix.
Covariances of \eqn{\mathbf{X}_{n \times 2, \cdots, k}}.}
}
\description{
Derives the \eqn{\boldsymbol{Omega}} matrix from linear regression model parameters.
}
\details{
The linear regression model is given by
\deqn{
    \mathbf{y}
    =
    \mathbf{X} \boldsymbol{\beta} + \boldsymbol{\varepsilon}
  }

where
\itemize{
\item \eqn{\mathbf{y}} is an \eqn{n \times 1} vector of observations
on the regressand variable
\item the data matrix \eqn{\mathbf{X}}
(also known as design matrix, model matrix or regressor matrix)
is an \eqn{n \times k} matrix of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation on the first column
\item \eqn{\boldsymbol{\beta}} is a \eqn{k \times 1} vector regression coefficients
\item \eqn{\boldsymbol{\varepsilon}} is the stochastic error term distributed around 0
with constant variance \eqn{\left( \sigma^2 \right)} across values of \eqn{\mathbf{X}}
}

Parameters of the linear regression model include
\itemize{
\item \eqn{\boldsymbol{\beta}} and
\item \eqn{\sigma^2}
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
