% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM2Data.R
\name{RAM2Data}
\alias{RAM2Data}
\title{Generate Data from a Multivariate Normal Distribution
Using the Reticular Action Model (RAM) Notation}
\usage{
RAM2Data(n, A, S, u, Filter, ...)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{Filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{...}{Additional arguments to pass to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\description{
Generates data from a multivariate normal distribution
using the Reticular Action Model (RAM) notation
with the \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}} function.
}
\details{
The Multivariate Normal Distribution is given by
\deqn{
    \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right)
  }

with location parameter

\deqn{
    \boldsymbol{\mu} \in \mathbf{R}^{k}
  }

and a positive definite covariance matrix

\deqn{
    \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} .
  }

The probability density function is given by

\deqn{
    f_{\mathbf{X}} \left( x_1, \cdots, x_k \right)
    =
    \frac{
      \exp
      \left[
        - \frac{1}{2}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)^{\mathsf{T}}
        \boldsymbol{\Sigma}^{-1}
        \left(
          \mathbf{x} - \boldsymbol{\mu}
        \right)
      \right]
      }{
        \sqrt{
          \left( 2 \pi \right)^k
          | \boldsymbol{\Sigma} |
        }
      }
    }

In this function,
the model-implied mean vector and variance-covariance matrix
are used as parameters to generate the data.

\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u}
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u}
  }
\deqn{
    \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{C}
    \mathbf{F}^{\mathsf{T}}
  }
}
\examples{
# This is an example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

n <- 50
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(Filter) <- c("y", "x", "e")
rownames(Filter) <- c("y", "x")
RAM2Data(n, A, S, u, Filter)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other data generation functions: 
\code{\link{Eq2Data}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generation functions}
\keyword{data}
