% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_num}
\alias{ram_num}
\title{Expectations from the Reticular Action Model (RAM) Matrices - Numeric}
\usage{
ram_num(A, S, u = NULL, filter)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix
\eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}
}
\description{
Expectations from the Reticular Action Model (RAM) Matrices - Numeric
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
u <- c(0.00, 0.50, 0.00)
filter <- diag(2)
filter <- cbind(filter, 0)
ram_num(A, S, u, filter)
}
\author{
Ivan Jacob Agaloos Pesigan
}
