% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_sym}
\alias{ram_sym}
\title{Expectations from the Reticular Action Model (RAM) Matrices - Symbolic}
\usage{
ram_sym(A, S, u = NULL, filter, simplify = FALSE)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix
\eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{simplify}{Logical.
Simplify the results.}
}
\description{
Expectations from the Reticular Action Model (RAM) Matrices - Symbolic
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigma2x", "sigma2e")
u <- c("alpha", "mux", 0)
filter <- diag(2)
filter <- cbind(filter, 0)
ram_sym(A, S, u, filter)
}
\author{
Ivan Jacob Agaloos Pesigan
}
