% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramR.R
\docType{package}
\name{ramR}
\alias{ramR}
\title{ramR: Reticular Action Model (RAM) Notation}
\description{
A collection of functions I use in studying
the Reticular Action Model (RAM) notation.
}
\section{ramR functions}{

\itemize{
\item \code{\link[=mutheta]{mutheta()}} - used to derive the model-implied mean vector
\item \code{\link[=Sigmatheta]{Sigmatheta()}} - used to derive the model0implied variance-covariance
}
}

\author{
Ivan Jacob Agaloos Pesigan
}
