% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S}
\alias{S}
\alias{S.default}
\alias{S.yac_symbol}
\title{Matrix of Symmetric Paths \eqn{\mathbf{S}}}
\usage{
S(A, C, ...)

\method{S}{default}(A, C, ...)

\method{S}{yac_symbol}(A, C, str = TRUE, ysym = TRUE, simplify = FALSE, tex = FALSE, ...)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{C}{\verb{t x t} numeric matrix \eqn{\mathbf{C}}.
Model-implied variance-covariance matrix.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Derives the matrix of symmetric paths (double-headed arrows) \eqn{\mathbf{S}}
using the Reticular Action Model (RAM) notation.
}
\details{
The matrix of symmetric paths (double-headed arrows) \eqn{\mathbf{S}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{S}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{C}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}} \\
    =
    \mathbf{E}^{-1}
    \mathbf{C}
    \left(
      \mathbf{E}^{-1}
    \right)^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{C}_{t \times t}} represents
the model-implied variance-covariance matrix, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
C <- matrix(
  data = c(
    1.25, 0.25, 1.00,
    0.25, 0.25, 0.00,
    1.00, 0.00, 1.00
  ),
  nrow = dim(A)[1]
)
colnames(A) <- rownames(A) <- c("y", "x", "e")
S(A, C)

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
C <- matrix(
  data = c(
    "sigmax2*beta^2+sigmae2", "sigmax2*beta", "sigmae2",
    "sigmax2*beta", "sigmax2", 0,
    "sigmae2", 0, "sigmae2"
  ),
  nrow = dim(A)[1]
)
S(Ryacas::ysym(A), C)
S(Ryacas::ysym(A), C, tex = TRUE)
S(Ryacas::ysym(A), C, ysym = FALSE)
S(Ryacas::ysym(A), C, str = FALSE)

beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1
S(Ryacas::ysym(A), C)
S(Ryacas::ysym(A), C, tex = TRUE)
S(Ryacas::ysym(A), C, ysym = FALSE)
S(Ryacas::ysym(A), C, str = FALSE)
eval(S(Ryacas::ysym(A), C, str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
