% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S_sym}
\alias{S_sym}
\title{Matrix of Symmetric Paths
\eqn{\mathbf{S}} - Symmetric}
\usage{
S_sym(A, C)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{C}{\verb{t x t} numeric matrix \eqn{\mathbf{C}}.
Model-implied variance-covariance matrix.}
}
\value{
Returns the matrix of symmetric paths (double-headed arrows)
\eqn{\mathbf{S}}
derived from the \code{A} and \code{C} matrices.
}
\description{
Derives the matrix of symmetric paths (double-headed arrows)
\eqn{\mathbf{S}}
using the Reticular Action Model (RAM) notation.
}
\details{
The matrix of symmetric paths (double-headed arrows)
\eqn{\mathbf{S}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{S}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{C}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}} \\
    =
    \mathbf{E}^{-1}
    \mathbf{C}
    \left(
      \mathbf{E}^{-1}
    \right)^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{C}_{t \times t}} represents
the model-implied variance-covariance matrix, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- "beta"
A[1, 3] <- 1
C <- matrix(
  data = c(
    "sigma2x*beta^2+sigma2e", "sigma2x*beta", "sigma2e",
    "beta*sigma2x", "sigma2x", 0,
    "sigma2e", 0, "sigma2e"
  ),
  nrow = dim(A)[1]
)
S_sym(A, C)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other S functions: 
\code{\link{S_num}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{S functions}
\keyword{S}
