% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u_sym}
\alias{u_sym}
\title{Vector of Mean Structure Parameters
\eqn{\mathbf{u}} - Symbolic}
\usage{
u_sym(A, v, simplify = FALSE)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{v}{vector of length \code{t} or \verb{t by 1} matrix.
Expected values.}

\item{simplify}{Logical.
Simplify the results.}
}
\value{
Returns the vector of mean structire parameters
\eqn{\mathbf{u}}.
}
\description{
Derives the vector of mean structure parameters
\eqn{\mathbf{u}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of mean structure parameters
\eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{u}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{v} \\
    =
    \mathbf{E}^{-1}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{v}_{t \times 1}} vector of expected values.
}
}
\examples{
# This is a symbolic example for the model
# y = alpha + beta * x + e

A <- matrixR::zeroes(3, 3)
A[1, ] <- c(0, "beta", 1)
v <- c("alpha+beta*mux", "mux", 0)
u_sym(A, v)
}
\seealso{
Other u functions: 
\code{\link{u_num}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{u functions}
\keyword{u}
