% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\alias{v.default}
\alias{v.yac_symbol}
\title{Vector of Expected Values \eqn{\mathbf{v}}}
\usage{
v(A, u, check = TRUE, ...)

\method{v}{default}(A, u, check = TRUE, ...)

\method{v}{yac_symbol}(
  A,
  u,
  check = TRUE,
  exe = TRUE,
  R = FALSE,
  format = "ysym",
  simplify = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{check}{Logical.
If \code{check = TRUE} do some preprocessing
with input matrices using \code{\link[=CheckRAMMatrices]{CheckRAMMatrices()}}.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE},
executes the resulting \code{yacas} expression.
If \code{exe = FALSE},
returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE},
the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
\eqn{\mathbf{v} = \mathbf{E} \mathbf{u}}
}
\description{
Derives the vector of expected values \eqn{\mathbf{v}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{v}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{E}
    \mathbf{u}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters
for the mean structure.
}
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
v(A, u)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
u <- c("alpha", "mux", 0)
v(Ryacas::ysym(A), u)
v(Ryacas::ysym(A), u, format = "str")
v(Ryacas::ysym(A), u, format = "tex")
v(Ryacas::ysym(A), u, R = TRUE)

# Assigning values to symbols

alpha <- 0
beta <- 1
mux <- 0.50

v(Ryacas::ysym(A), u)
v(Ryacas::ysym(A), u, format = "str")
v(Ryacas::ysym(A), u, format = "tex")
v(Ryacas::ysym(A), u, R = TRUE)
eval(v(Ryacas::ysym(A), u, R = TRUE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{RAMScaled}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
