% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_num}
\alias{v_num}
\title{Vector of Expected Values
\eqn{\mathbf{v}} - Numeric}
\usage{
v_num(A, u)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}
}
\value{
Returns the vector of expected values
\eqn{\mathbf{v}}.
}
\description{
Derives the vector of expected values
\eqn{\mathbf{v}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values
\eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{v}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{E}
    \mathbf{u}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters for the mean structure.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
v_num(A, u)
}
\seealso{
Other v functions: 
\code{\link{v_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{v functions}
\keyword{v}
