% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YacExe.R
\name{YacExe}
\alias{YacExe}
\title{Execute Yacas Expression using Ryacas}
\usage{
YacExe(expr, str = TRUE, ysym = TRUE, simplify = FALSE, tex = FALSE)
}
\arguments{
\item{expr}{Character string. Yacas expresion.}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Execute \code{yacas} expressions using \code{Ryacas}.
}
\examples{
A <- Ryacas::ysym(
  matrix(c("a", "c", "b", "d"), ncol = 2)
)
expr <- paste0("Determinant(", A, ")")
YacExe(expr)
YacExe(expr, str = TRUE, ysym = FALSE)
YacExe(expr, str = TRUE, tex = TRUE)
YacExe(expr, str = FALSE)

a <- 1
b <- 2
c <- 3
d <- 4
eval(YacExe(expr, str = FALSE))

A <- matrix(c(a, c, b, d), ncol = 2)
det(A)
}
\references{
\href{http://www.yacas.org/}{YACAS - Yet Another Computer Algebra System}

\href{http://r-cas.github.io/ryacas/}{Ryacas}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Yacas functions}
\keyword{yac}
