% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\alias{E.default}
\alias{E.yac_symbol}
\title{Matrix of Total Effects \eqn{\mathbf{E}}}
\usage{
E(A, ...)

\method{E}{default}(A, ...)

\method{E}{yac_symbol}(A, exe = TRUE, str = TRUE, ysym = TRUE, simplify = FALSE, tex = FALSE, ...)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE}, executes the resulting \code{yacas} expression.
If \code{exe = FALSE}, returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE}, the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{str}{Logical.
If \code{str = TRUE}, returns symbolic result as string.
If \code{str = FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{ysym = TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{ysym = FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical.
If \code{tex = TRUE} returns results as latex math if \code{str = TRUE}.
If \code{tex = FALSE}, returns symbolic result as string if \code{str = TRUE}.}
}
\value{
\eqn{\mathbf{E} = \left( \mathbf{I} - \mathbf{A} \right)^{-1}}
}
\description{
Derives the matrix of total effects \eqn{\mathbf{E}}.
}
\details{
The matrix of total effects \eqn{\mathbf{E}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{E}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
E(A)

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
E(Ryacas::ysym(A))
E(Ryacas::ysym(A), tex = TRUE)
E(Ryacas::ysym(A), ysym = FALSE)
E(Ryacas::ysym(A), str = FALSE)

beta <- 1
E(Ryacas::ysym(A))
E(Ryacas::ysym(A), tex = TRUE)
E(Ryacas::ysym(A), ysym = FALSE)
E(Ryacas::ysym(A), str = FALSE)
eval(E(Ryacas::ysym(A), str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
