% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{Matrix of Total Effects \eqn{\mathbf{E}}}
\usage{
E(A)
}
\arguments{
\item{A}{\verb{t x t} numeric matrix
\eqn{\mathbf{A}_{t \times t}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}
}
\description{
Derives the matrix of total effects \eqn{\mathbf{E}}.
}
\details{
The matrix of total effects \eqn{\mathbf{E}} is given by
\deqn{
    \mathbf{E}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
  }
}
\seealso{
Other SEM notation functions: 
\code{\link{M}()},
\code{\link{Omega}()},
\code{\link{Sigmatheta}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
