% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq2data.R
\name{eq2data}
\alias{eq2data}
\title{Equations to Sample Data}
\usage{
eq2data(model, n, ...)
}
\arguments{
\item{model}{Character string. Input model. See Details.}

\item{n}{Integer.
Sample size.}

\item{...}{Additional arguments to pass to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\description{
Generates data from a multivariate normal distribution
from an input model with specified values.
}
\details{
The input model is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{VARIABLE1 OPERATION VARIABLE2 VALUE}

The associations are defined by the following operations

\describe{
\item{by}{\code{VARIABLE1} measured \strong{by} \code{VARIABLE2}}
\item{on}{\code{VARIABLE1} regressed \strong{on} \code{VARIABLE2}}
\item{with}{\code{VARIABLE1} covarying \strong{with} \code{VARIABLE2}}
\item{on 1}{\code{VARIABLE1} regressed \strong{on 1} for mean structure}
}

\strong{Each line should end with a semicolon (\verb{;}).}
}

\section{Value}{

Each parameter should have a numeric value.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
model <- "
  # VARIABLE1 OPERATION VARIABLE2 VALUE
  e           by        y         1.00;
  y           on        x         1.00;
  e           with      e         0.25;
  x           with      x         0.25;
  y           on        1         0.00;
  x           on        1         0.50
"
eq2data(model, n = 100)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
