% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq2Expectations.R
\name{Eq2Expectations}
\alias{Eq2Expectations}
\title{Equations to Expectations}
\usage{
Eq2Expectations(
  eq,
  par = FALSE,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE
)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{par}{Logical.
If \code{par = TRUE}, use \code{par.index} as labels.
If \code{par = FALSE}, use \code{label} as labels.}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Converts equations to expectations.
}
\details{
The input is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{lhs operation rhs label}

The associations are defined by the following operations

\describe{
\item{by}{\verb{left-hand side} measured \strong{by} \verb{right-hand side}}
\item{on}{\verb{left-hand side} regressed \strong{on} \verb{right-hand side}}
\item{with}{\verb{left-hand side} covarying \strong{with} \verb{right-hand side}}
\item{on 1}{\verb{left-hand side} regressed \strong{on 1} for mean structure}
}
}

\section{label}{

Each parameter should be labeled.
The \code{label} should be a number for fixed parameters
and a character string for free parameters.
Equality contraints can be imposed by using the same label.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
# Numerical -------------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"

# Symbolic -------------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2Expectations(eq, par = FALSE)

# Expressions using `label`

beta <- 1
sigmae2 <- 1
sigmax2 <- 0.25
alpha <- 0
mux <- 0.50
Exp <- Eq2Expectations(eq, par = FALSE, str = FALSE)
eval(Exp$M)
eval(Exp$g)

# Expressions using `par.index`

p <- c(beta, sigmae2, sigmax2, alpha, mux)
p1 <- p[1]
p2 <- p[2]
p3 <- p[3]
p4 <- p[4]
p5 <- p[5]
Exp <- Eq2Expectations(eq, par = TRUE, str = FALSE)
eval(Exp$M)
eval(Exp$g)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
