% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq2RAM.R
\name{Eq2RAM}
\alias{Eq2RAM}
\title{Equations to RAM Matrices}
\usage{
Eq2RAM(eq, par = FALSE)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{par}{Logical.
If \code{par = TRUE}, use \code{par.index} as labels.
If \code{par = FALSE}, use \code{par.label} as labels.}
}
\value{
Returns a list with the following elements

\describe{
\item{par.table}{Parameter table.}
\item{variables}{Variable names.}
\item{g.variables}{Variable names of observed variables.}
\item{h.variables}{Variable names of latent variables.}
\item{A}{
\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.
}
\item{S}{
\verb{t by t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.
}
\item{u}{
\verb{t by 1} matrix \eqn{\mathbf{u}} of mean structure parameters.
}
\item{Filter}{
\verb{p by t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.
}
}
}
\description{
Converts model equations to RAM matrices.
}
\details{
The input is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{lhs <space> op <space> rhs <space> par.label <\\n> or <;>}

\describe{
\item{lhs}{
is the variable on the \strong{left-hand side},
}
\item{rhs}{
is the variable on the \strong{right-hand side},
}
\item{op}{
is the \strong{operation} between \code{lhs} and \code{rhs},
}
\item{par.label}{
is the column of \strong{parameter label},
}
\item{\\n or ;}{
are \strong{line breaks}.
\strong{Each line should end with a line break.}
}
}
}

\section{Operations}{

The associations are defined by the following operations

\describe{
\item{by}{
\verb{left-hand side} measured \strong{by} \verb{right-hand side},
}
\item{on}{
\verb{left-hand side} regressed \strong{on} \verb{right-hand side},
}
\item{with}{
\verb{left-hand side} covarying \strong{with} \verb{right-hand side},
}
\item{on 1}{
\verb{left-hand side} regressed \strong{on 1} for mean structure.
}
}
}

\section{par.label}{

Each parameter should be labeled.
The \code{par.label} should be a number for fixed parameters
and a character string for free parameters.
Equality contraints can be imposed by using the same \code{par.label}.
}

\section{Line breaks}{

The characters \verb{\\n} and \verb{;} can be used as line breaks.
\strong{Each line should end with a line break.}
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
# Numeric ----------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2RAM(eq)

# Symbolic ----------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2RAM(eq)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other eq functions: 
\code{\link{Eq2Expectations}()},
\code{\link{EqParse}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{eq functions}
\keyword{eq}
