% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq2RAM.R
\name{Eq2RAM}
\alias{Eq2RAM}
\title{Equations to RAM Matrices}
\usage{
Eq2RAM(eq, par = FALSE)
}
\arguments{
\item{eq}{Character string. Equations. See Details.}

\item{par}{Logical.
If \code{par = TRUE}, use \code{par.index} as labels.
If \code{par = FALSE}, use \code{label} as labels.}
}
\description{
Converts model equations to RAM matrices.
}
\details{
The input is a character string
that specifies the associations between the variables.
}
\section{Syntax}{

Each line should follow the syntax below

\verb{lhs operation rhs label}

The associations are defined by the following operations

\describe{
\item{by}{\verb{left-hand side} measured \strong{by} \verb{right-hand side}}
\item{on}{\verb{left-hand side} regressed \strong{on} \verb{right-hand side}}
\item{with}{\verb{left-hand side} covarying \strong{with} \verb{right-hand side}}
\item{on 1}{\verb{left-hand side} regressed \strong{on 1} for mean structure}
}
}

\section{label}{

Each parameter should be labeled.
The \code{label} should be a number for fixed parameters
and a character string for free parameters.
Equality contraints can be imposed by using the same label.
}

\section{Comments}{

Comments can be written after a hash (\verb{#}) sign.
}

\examples{
# Numeric ----------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2RAM(eq, par = FALSE)
Eq2RAM(eq, par = TRUE)

# Symbolic ----------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2RAM(eq, par = FALSE)
Eq2RAM(eq, par = TRUE)
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
