% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\alias{g.default}
\alias{g.yac_symbol}
\title{Vector of Expected Values of Observed Variables \eqn{\mathbf{g}}}
\usage{
g(A, u, Filter = NULL, ...)

\method{g}{default}(
  A,
  u,
  Filter = NULL,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)

\method{g}{yac_symbol}(
  A,
  u,
  Filter = NULL,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{Filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Derives the vector of expected values of observed variables \eqn{\mathbf{g}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values of observed variables \eqn{\mathbf{g}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{g}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{u} \\
    =
    \mathbf{F}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters for the mean structurem,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{p + q} .
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(A) <- rownames(A) <- c("y", "x", "e")
g(A, u, Filter)

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
u <- c("alpha", "mux", 0)
g(Ryacas::ysym(A), u, Filter)
g(Ryacas::ysym(A), u, Filter, tex = TRUE)
g(Ryacas::ysym(A), u, Filter, ysym = FALSE)
g(Ryacas::ysym(A), u, Filter, str = FALSE)

alpha <- 0
beta <- 1
mux <- 0.50
g(Ryacas::ysym(A), u, Filter)
g(Ryacas::ysym(A), u, Filter, tex = TRUE)
g(Ryacas::ysym(A), u, Filter, ysym = FALSE)
g(Ryacas::ysym(A), u, Filter, str = FALSE)
eval(g(Ryacas::ysym(A), u, Filter, str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
