% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAMR.R
\name{getAMR}
\alias{getAMR}
\title{Search for aberrantly methylated regions}
\usage{
getAMR(
  data.ranges,
  data.samples = NULL,
  data.coverage = NULL,
  transform = c("identity", "linear"),
  exclude.range = NULL,
  compute = c("IQR", "beta+binom"),
  compute.estimate = c("mom", "amle", "nmle"),
  compute.weights = c("equal", "logInvDist", "sqrtInvDist", "invDist"),
  combine = c("threshold", "comb-p"),
  combine.threshold = ifelse(compute == "IQR", 5, 0.001),
  combine.window = 300,
  combine.min.cpgs = 7,
  combine.min.width = 1,
  combine.ignore.strand = FALSE,
  ncores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data.ranges}{A `GRanges` object with genomic locations and
corresponding beta values included as metadata.}

\item{data.samples}{A character vector with sample names (a subset of
metadata column names). If `NULL` (the default), then all samples (metadata
columns) are included in the analysis.}

\item{data.coverage}{description}

\item{transform}{description}

\item{exclude.range}{A numeric vector of length two. If \emph{not} `NULL`
(the default), all `data.ranges` genomic locations with their median
methylation beta value within the `exclude.range` interval are filtered out.}

\item{compute}{A character scalar: when ramr.method is "IQR" (the
default), the filtering based on interquantile range is used (`iqr.cutoff`
value is then used as a threshold). When "beta", "wbeta" or "beinf" -
filtering based on fitting non-weighted (`EnvStats::ebeta`), weighted
(`ExtDist::eBeta`) or zero-and-one inflated (`gamlss.dist::BEINF`)
beta distributions, respectively, is used, and `pval.cutoff` or `qval.cutoff`
(if not `NULL`) is used as a threshold. For "wbeta", weights directly
correlate with bin contents (number of values per bin) and inversly - with
the distances from the median value, thus narrowing the estimated
distribution and emphasizing outliers.}

\item{compute.estimate}{description}

\item{compute.weights}{A single string for the weights assigned to individual
observations (beta values) and used to compute weighted
means and variance as described in "Compute" section below. Four available
weighing schemes result in different sensitivity of outlier detection and
rate of false positive (FP) findings: "equal" (the default) is the least
sensitive and gives least number of FPs, while "invDist" is the most
sensitive but may result in a very high number of FPs especially when
`combine.threshold` is too high (1e-3 or higher).
"logInvDist" is recommended when one desires a balance between relatively
low type I error rate and higher
detection sensitivity for both unique and non-unique AMRs.

If "equal", all weights are equal to 1 (\eqn{w_i=1}).
Otherwise, weights of observations inversely depend on their distance from
the median (thus emphasizing outliers) and are calculated using the
following formulas:
\deqn{\text{"logInvDist": } w_i = \log{\frac{1}{|M-x_i|+\epsilon}}}
\deqn{\text{"sqrtInvDist": } w_i = \sqrt{\frac{1}{|M-x_i|+\epsilon}}}
\deqn{\text{"invDist": } w_i = \frac{1}{|M-x_i|+\epsilon}}
where \eqn{x_i} is a beta value for i-th sample, \eqn{M} is a median of
all beta values at this genomic location, and \eqn{\epsilon} is a very
small number (= FLT_EPSILON \eqn{\approx} 1.192093e-07).}

\item{combine}{description}

\item{combine.threshold}{A single integer >= 1. Methylation beta values differing
from the median value by more than `iqr.cutoff` interquartile ranges are
considered to be significant (the default: 5).
param pval.cutoff A numeric scalar (the default: 5e-2). Bonferroni
correction of `pval.cutoff` by the length of the `data.samples` object is
used to calculate `qval.cutoff` if the latter is `NULL`.
param qval.cutoff A numeric scalar. Used as a threshold for filtering based
on fitting non-weighted or weighted beta distributions: all p-values lower
than `qval.cutoff` are considered to be significant. If `NULL` (the default),
it is calculated using `pval.cutoff`}

\item{combine.window}{A positive integer. All significant (survived the
filtering stage) `data.ranges` genomic locations within this distance will be
merged to create AMRs (the default: 300).}

\item{combine.min.cpgs}{A single integer >= 1. All AMRs containing less than
`min.cpgs` significant genomic locations are filtered out (the default: 7).}

\item{combine.min.width}{A single integer >= 1 (the default). Only AMRs with the
width of at least `min.width` are returned.}

\item{combine.ignore.strand}{description}

\item{ncores}{A single integer >= 1. Number of processes for parallel
computation (the default: all but one cores). Results of parallel processing
are fully reproducible when the same seed is used (thanks to doRNG).}

\item{verbose}{Boolean to report progress and timings (default: TRUE).}
}
\value{
The output is a `GRanges` object that contains all the aberrantly
methylated regions (AMRs) for all `data.samples` samples in `data.ranges`
object. The following metadata columns may be present:
\itemize{
  \item `revmap` -- integer list of significant CpGs (`data.ranges` genomic
  locations) that are included in this AMR region
  \item `ncpg` -- number of significant CpGs within this AMR region
  \item `sample` -- contains an identifier of a sample to which
  corresponding AMR belongs
  \item `dbeta` -- average deviation of beta values for significant CpGs from
  their corresponding median values
  \item `pval` -- geometric mean of p-values for significant CpGs
  \item `xiqr` -- average IQR-normalised deviation of beta values for
  significant CpGs from their corresponding median values
}
}
\description{
`getAMR` returns a `GRanges` object with aberrantly methylated
regions (AMRs / epimutations) for all samples in a data set.
}
\details{
In the provided data set, `getAMR` finds stretches of outlier beta values
to identify rare long-range methylation
aberrations (epimutations) in one or several samples. Other methods for
differential methylation analysis rely on between-group comparisons ---
`getAMR` performs this comparison within-group, which is not only faster,
but also more sensitive. The logic of computations is described below.

\subsection{Compute}{
There are only two outlier detection methods that are supported at the
moment:

\describe{
  \item{"IQR"}{
    When `compute=="IQR"`, for every genomic location (CpG) in
    `data.ranges` the IQR-normalized deviation from the median value is
    calculated using the following formula:
    \deqn{xIQR_i=\frac{x_i-M}{IQR}}
    where \eqn{x_i} is a beta value for i-th sample,
    \eqn{M} and \eqn{IQR} are a median and an interquartile range
    of all beta values at this genomic location, respectively.
  }
  \item{"beta+binom"}{
    When `compute=="beta+binom"`, for every genomic location (CpG),
    `getAMR` will estimate the probability of each beta value to occur.
    For all beta values inside the open \eqn{(0,1)} interval, beta
    distribution is used. For all \eqn{\{0;1\}} endpoint (extreme) values
    where beta distribution is not defined, binomial probability is
    calculated using coverage data (supplied using `data.coverage`
    parameter).

    Alpha \eqn{{\alpha}} and beta \eqn{{\beta}} parameters of beta
    distribution can be estimated using one of the following methods:

    \enumerate{
      \item{
        Method of moments (`compute.estimate="mom"`) based on
        (both optionally weighted) mean and unbiased variance
        \deqn{\text{sample mean} = \bar{x} = \frac{ \sum\limits_{i=1}^n w_i x_i}{\sum\limits_{i=1}^n w_i}}
        \deqn{\text{unbiased variance} =\bar{v} = \frac {\sum\limits_{i=1}^N w_i (x_i - \bar{x})^2} {\sum_{i=1}^N w_i - (\sum_{i=1}^N w_i^2 / \sum_{i=1}^N w_i)} }
        where \eqn{x_i} and \eqn{w_i} are a beta value and its reliability
        weight for i-th sample.
        \deqn{\hat{\alpha} = \bar{x} \left(\frac{\bar{x} (1 - \bar{x})}{\bar{v}} - 1 \right)}
        \deqn{\hat{\beta} = (1-\bar{x}) \left(\frac{\bar{x} (1 - \bar{x})}{\bar{v}} - 1 \right)}
        For more details, visit
        \href{https://en.wikipedia.org/wiki/Weighted_arithmetic_mean#Mathematical_definition}{Weighted arithmetic mean},
        \href{https://en.wikipedia.org/wiki/Weighted_arithmetic_mean#Reliability_weights}{Weighted variance with reliability weights},
        and \href{https://en.wikipedia.org/wiki/Beta_distribution#Method_of_moments}{Method of moments for beta distribution}
      }
      \item{
        Approximate maximum likelihood (`compute.estimate="amle"`) based on
        (both optionally weighted) geometric means of \eqn{x} and \eqn{(1-x)}
        \deqn{ \hat{G}_x = \left(\prod_{i=1}^n x_i^{w_i}\right)^{1 / \sum_{i=1}^n w_i} }
        \deqn{ \hat{G}_{(1-x)} = \left(\prod_{i=1}^n {(1-x_i)}^{w_i}\right)^{1 / \sum_{i=1}^n w_i} }
        where \eqn{x_i} and \eqn{w_i} are a beta value and its reliability
        weight for i-th sample.
        \deqn{ \hat{\alpha}\approx \tfrac{1}{2} + \frac{\hat{G}_{x}}{2(1-\hat{G}_x-\hat{G}_{(1-x)})} }
        \deqn{ \hat{\beta}\approx \tfrac{1}{2} + \frac{\hat{G}_{(1-x)}}{2(1-\hat{G}_x-\hat{G}_{(1-x)})} }
        For more details, visit
        \href{https://en.wikipedia.org/wiki/Weighted_geometric_mean}{Weighted geometric mean},
        and \href{https://en.wikipedia.org/wiki/Beta_distribution#Two_unknown_parameters_2}{Maximum likelihood for beta distribution}
      }
      \item{
        And a numerical maximum likelihood (`compute.estimate="nmle"`) which
        is yet to be implemented.
      }
    }



  }
}

, and all CpGs with such normalized deviation not smaller
    than the `iqr.cutoff` are retained.

 For
`ramr.method %in% c("beta", "wbeta", "beinf")`: parameters of beta
distribution are estimated by means of `EnvStats::ebeta` (beta distribution),
`ExtDist::eBeta` (weighted beta destribution), or `gamlss.dist::BEINF` (zero
and one inflated beta distribution) functions, respectively. These
parameters are then used to calculate the probability values, followed by the
filtering when all CpGs with p-values not greater than `qval.cutoff` are
retained. Another filtering is then performed to exclude all CpGs within
`exclude.range`. Next, the retained (significant) CpGs are merged within
the window of `merge.window`, and final filtering is applied to AMR genomic
ranges (by `min.cpgs` and `min.width`).
}

\subsection{Combine}{
this and that
}
}
\examples{
  data(ramr)
  getAMR(data.ranges=ramr.data, data.samples=ramr.samples,
         compute="beta+binom", compute.estimate="amle",
         combine.min.cpgs=5, combine.window=1000, combine.threshold=1e-3)
}
\seealso{
\code{\link{plotAMR}} for plotting AMRs, \code{\link{getUniverse}}
for info on enrichment analysis, \code{\link{simulateAMR}} and
\code{\link{simulateData}} for the generation of simulated test data sets,
and `ramr` vignettes for the description of usage and sample data.
}
