% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.R
\name{M}
\alias{M}
\alias{M.default}
\alias{M.yac_symbol}
\title{Matrix of Covariance Expectations of Observed Variables
\eqn{\mathbf{M}}}
\usage{
M(A, S, Filter, ...)

\method{M}{default}(A, S, Filter = NULL, ...)

\method{M}{yac_symbol}(
  A,
  S,
  Filter = NULL,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\verb{t x t} numeric matrix \eqn{\mathbf{S}}.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{Filter}{\verb{p x t} numeric matrix
\eqn{\mathbf{F}}.
Filter matrix used to select observed variables.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Derives the matrix of covariance expectations
of observed variables \eqn{\mathbf{M}}.
}
\details{
The matrix of covariance expectations
for given variables \eqn{\mathbf{M}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{E}
    \mathbf{S}
    \mathbf{E}^{\mathsf{T}}
    \mathbf{F}^{\mathsf{T}} \\
    =
    \mathbf{F}
    \mathbf{C}
    \mathbf{F}^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{S}_{t \times t}} represents symmetric paths
(double-headed arrows),
such as variances and covariances,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\mathbf{F}_{p \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{p} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{t} number of observed and latent variables,
that is \eqn{p + q} .
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(Filter) <- c("y", "x", "e")
M(A, S, Filter)

# Symbolic ----------------------------------------------------------
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigmax2", "sigmae2")
M(Ryacas::ysym(A), S, Filter)
M(Ryacas::ysym(A), S, Filter, tex = TRUE)
M(Ryacas::ysym(A), S, Filter, ysym = FALSE)
M(Ryacas::ysym(A), S, Filter, str = FALSE)

beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1
M(Ryacas::ysym(A), S, Filter)
M(Ryacas::ysym(A), S, Filter, tex = TRUE)
M(Ryacas::ysym(A), S, Filter, ysym = FALSE)
M(Ryacas::ysym(A), S, Filter, str = FALSE)
eval(M(Ryacas::ysym(A), S, Filter, str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
