% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Mean Structure Vector \eqn{\mathbf{m}}}
\usage{
m(mutheta, A, filter)
}
\arguments{
\item{mutheta}{\verb{t x 1} numeric vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)_{t \times 1}} .
Model-implied mean vector.}

\item{A}{\verb{t x t} numeric matrix
\eqn{\mathbf{A}_{t \times t}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{filter}{\verb{j x t} numeric matrix
\eqn{\mathbf{F}_{j \times t}}.
Filter matrix used to select variables.}
}
\value{
Returns the mean structure vector \eqn{\mathbf{m}}
derived from the \code{mutheta}, \code{A}, and \code{filter} matrices.
}
\description{
Derives the mean structure vector \eqn{\mathbf{m}}
using the Reticular Action Model (RAM) notation.
}
\details{
The mean structure vector \eqn{\mathbf{m}}
as a function of Reticular Action Model (RAM) matrices is given by

\deqn{
    \mathbf{m}
    =
    \mathbf{F}
    \left( \mathbf{I} - \mathbf{A} \right)
    \mathbf{F}^{\mathsf{T}}
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{F}_{j \times t}} represents the filter matrix
used to select the observed variables,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix,
\item \eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
is the \eqn{t \times 1} model-implied mean vector
\item \eqn{j} number of observed variables,
\item \eqn{k} number of latent variables, and
\item \eqn{t} number of observed and latent variables, that is \eqn{j + k} .
}
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules
for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald}
(pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{Omega}()},
\code{\link{Sigmatheta}()},
\code{\link{mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
