% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAMR.R
\name{plotAMR}
\alias{plotAMR}
\title{Plot aberrantly methylated regions}
\usage{
plotAMR(
  data.ranges,
  data.samples,
  ramr.ranges,
  highlight = NULL,
  title = NULL,
  window = 300
)
}
\arguments{
\item{data.ranges}{A `GRanges` object with genomic locations and corresponding
beta values included as metadata.}

\item{data.samples}{A character vector with sample names (a subset of metadata
column names) to be included in the plot.}

\item{ramr.ranges}{An output of `getAMR` - a `GRanges` object that contain
aberrantly methylated regions (AMRs).}

\item{highlight}{An optional list of samples to highlight. If NULL (default),
will contain sample IDs from the `sample` metadata column of `ramr.ranges` object.}

\item{title}{An optional title for the plot. If NULL (default), plot title is
set to a genomic location of particular AMR.}

\item{window}{An optional integer constant to expand genomic ranges of the
`ramr.ranges` object.}
}
\value{
The output is a list of `ggplot` objects.
}
\description{
`plotAMR` uses `ggplot2` to visualize aberrantly methylated regions (AMRs)
at particular genomic locations.
}
\details{
For every non-overapping genomic location from `ramr.ranges` object, `plotAMR`
plots and outputs a line graph of methylation beta values taken from `data.ranges`
for all samples from `data.samples`. Samples bearing significantly different
methylation profiles ('sample' column of `ramr.ranges` object) are highlighted.
}
\examples{
\dontrun{
  plotAMR(ramr.data, ramr.samples, ramr.tp.unique[1])

  # library(gridExtra)
  do.call("grid.arrange", c(plotAMR(ramr.data, ramr.samples, ramr.tp.nonunique), ncol=2))
}
}
\seealso{
\code{\link{getAMR}} for identification of AMRs, \code{\link{getUniverse}} or README for
info on enrichment analysis
}
