% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramr-data.R
\docType{data}
\name{ramr.data}
\alias{ramr.data}
\alias{ramr.samples}
\alias{ramr.tp.unique}
\alias{ramr.tp.nonunique}
\title{Simulated Illumina HumanMethylation 450k dataset with 3000 CpGs and 100 samples}
\format{Objects of class \code{"GRanges"} (\code{"ramr.data, ramr.tp.unique, ramr.tp.nonunique"})
and \code{"character"} (\code{"ramr.samples"}).}
\usage{
data(ramr)
}
\description{
Data was simulated using GSE51032 dataset as described in the reference.
Current dataset (\code{"ramr.data"}) contains beta values for 10000 CpGs
and 100 samples (\code{"ramr.samples"}), and carries 6 unique
(\code{"ramr.tp.unique"}) and 15 non-unique (\code{"ramr.tp.nonunique"})
true positive AMRs containing at least 10 CpGs with their beta values
increased/decreased by 0.5
}
\examples{
\dontrun{
  data(ramr)
  amrs <- getAMR(ramr.data, ramr.samples, ramr.method="beta", min.cpgs=5,
                 merge.window=1000, qval.cutoff=1e-3)
  plotAMR(ramr.data, ramr.samples, amrs[1])
  plotAMR(ramr.data, ramr.samples, ramr.tp.nonunique[4], highlight=c("sample7","sample8","sample9"))
}
}
\references{
Nikolaienko et al. (2020) bioRxiv
(\href{https://www.biorxiv.org/content/10.1101/draft}{bioRxiv})
}
\keyword{datasets}
