% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramR.R
\docType{package}
\name{ramR}
\alias{ramR}
\title{ramR: Reticular Action Model (RAM) Notation}
\description{
A collection of utility functions using
the Reticular Action Model (RAM) notation.
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\author{
Ivan Jacob Agaloos Pesigan
}
