% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S}
\alias{S}
\alias{S.default}
\alias{S.yac_symbol}
\title{Matrix of Symmetric Paths \eqn{\mathbf{S}}}
\usage{
S(A, C, ...)

\method{S}{default}(A, C, ...)

\method{S}{yac_symbol}(A, C, exe = TRUE, R = FALSE, format = "ysym", simplify = FALSE, ...)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{C}{\verb{t by t} numeric matrix \eqn{\mathbf{C}}.
Model-implied variance-covariance matrix.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE},
executes the resulting \code{yacas} expression.
If \code{exe = FALSE},
returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE},
the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
\eqn{\mathbf{S} = \mathbf{E}^{-1} \mathbf{C}
  \left( \mathbf{E}^{-1} \right)^{\mathsf{T}}}
}
\description{
Derives the matrix of symmetric paths (double-headed arrows)
\eqn{\mathbf{S}} using the Reticular Action Model (RAM) notation.
}
\details{
The matrix of symmetric paths (double-headed arrows) \eqn{\mathbf{S}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{S}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{C}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}} \\
    =
    \mathbf{E}^{-1}
    \mathbf{C}
    \left(
      \mathbf{E}^{-1}
    \right)^{\mathsf{T}}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{C}_{t \times t}} represents
the model-implied variance-covariance matrix, and
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix.
}
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
C <- matrix(
  data = c(
    1.25, 0.25, 1.00,
    0.25, 0.25, 0.00,
    1.00, 0.00, 1.00
  ),
  nrow = dim(A)[1]
)
colnames(A) <- rownames(A) <- c("y", "x", "e")
S(A, C)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
C <- matrix(
  data = c(
    "sigmax2*beta^2+sigmae2", "sigmax2*beta", "sigmae2",
    "sigmax2*beta", "sigmax2", 0,
    "sigmae2", 0, "sigmae2"
  ),
  nrow = dim(A)[1]
)
S(Ryacas::ysym(A), C, R = FALSE, format = "ysym")
S(Ryacas::ysym(A), C, R = FALSE, format = "str")
S(Ryacas::ysym(A), C, R = FALSE, format = "tex")
S(Ryacas::ysym(A), C, R = TRUE)

# Assigning values to symbols

beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1
S(Ryacas::ysym(A), C, R = FALSE, format = "ysym")
S(Ryacas::ysym(A), C, R = FALSE, format = "str")
S(Ryacas::ysym(A), C, R = FALSE, format = "tex")
S(Ryacas::ysym(A), C, R = TRUE)
eval(S(Ryacas::ysym(A), C, R = TRUE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{g}()},
\code{\link{u}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
