% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateAMR.R
\name{simulateAMR}
\alias{simulateAMR}
\title{Simulate a set of aberrantly methylated regions}
\usage{
simulateAMR(
  template.ranges,
  nsamples,
  exclude.ranges = NULL,
  regions.per.sample = 1,
  samples.per.region = 1,
  sample.names = NULL,
  merge.window = 300,
  min.cpgs = 7,
  max.cpgs = Inf,
  min.width = 1,
  dbeta = 0.25
)
}
\arguments{
\item{merge.window}{A positive integer. All significant (survived the
filtering stage) `data.ranges` genomic locations within this distance will be
merged to create AMRs.}

\item{min.cpgs}{A single integer >= 1. All AMRs containing less than
`min.cpgs` significant genomic locations are filtered out.}

\item{min.width}{A single integer >= 1 (the default). Only AMRs with the
width of at least `min.width` are returned.}

\item{data.ranges}{A `GRanges` object with genomic locations and
corresponding beta values included as metadata.}

\item{data.samples}{A character vector with sample names (a subset of
metadata column names).}

\item{ramr.method}{A character scalar: when ramr.method is "IQR" (the
default), the filtering based on interquantile range is used (`iqr.cutoff`
value is then used as a threshold). When "beta" or "wbeta" - filtering based
on fitting non-weighted (EnvStats::ebeta) or weighted (ExtDist::eBeta) beta
distributions, respectively, is used, and `pval.cutoff` or `qval.cutoff` (if
not `NULL`) is used as a threshold. For "wbeta", weights directly correlate
with bin contents (number of values per bin) and inversly - with the
distances from the median value, thus narrowing the estimated distribution
and emphasizing outliers.}

\item{iqr.cutoff}{A single integer >= 1. Methylation beta values differing
from the median value by more than `iqr.cutoff` interquartile ranges are
considered to be significant.}

\item{pval.cutoff}{A numeric scalar. Bonferroni correction of `pval.cutoff`
by the length of the `data.ranges` object is used to calculate `qval.cutoff`
if the latter is `NULL`.}

\item{qval.cutoff}{A numeric scalar. Used as a threshold for filtering based
on fitting non-weighted or weighted beta distributions: all p-values lower
than `qval.cutoff` are considered to be significant. If `NULL` (the default),
it is calculated using `pval.cutoff`}

\item{exclude.range}{A numeric vector of length two. If not `NULL`, all
`data.ranges` genomic locations with their median methylation beta value
within the `exclude.range` interval are filtered out.}

\item{cores}{A single integer >= 1. Number of processes for parallel
computation.}

\item{...}{Further arguments to be passed to `EnvStats::ebeta` or
`ExtDist::eBeta` functions.}
}
\value{
The output is a `GRanges` object that contain all the aberrantly
methylated regions (AMRs) for all `data.samples` samples in `data.ranges`
object. The `sample` metadata column contains an identifier of a sample to
which corresponding AMR belongs to.
}
\description{
`simulateAMR` returns a `GRanges` object with all the aberrantly methylated
regions (AMRs) for all samples in a dataset.
}
\details{
In the provided dataset `getAMR` compares methylation beta values of each
sample with other samples to identify rare long-range methylation
aberrations.
}
\examples{
  data(ramr)
  getAMR(ramr.data, ramr.samples, ramr.method="beta",
         min.cpgs=5, merge.window=1000, qval.cutoff=1e-3, cores=2)
}
\seealso{
\code{\link{plotAMR}} for plotting AMRs, \code{\link{getUniverse}}
for info on enrichment analysis
}
