% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\alias{u.default}
\alias{u.yac_symbol}
\title{Vector of Mean Structure Parameters \eqn{\mathbf{u}}}
\usage{
u(A, v, ...)

\method{u}{default}(A, v, ...)

\method{u}{yac_symbol}(A, v, str = TRUE, ysym = TRUE, simplify = FALSE, tex = FALSE, ...)
}
\arguments{
\item{A}{\verb{t x t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{v}{vector of length \code{t} or \verb{t by 1} matrix.
Expected values.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, returns symbolic result as string.
If \code{FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical. Return results as latex if \code{str = TRUE}.}
}
\description{
Derives the vector of mean structure parameters \eqn{\mathbf{u}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of mean structure parameters \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{u}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{v} \\
    =
    \mathbf{E}^{-1}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{v}_{t \times 1}} vector of expected values.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
v <- c(0.50, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
u(A, v)

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
v <- c("alpha+beta*mux", "mux", 0)
# u(A, v, sym = TRUE)
# u(A, v, sym = TRUE, str = FALSE)
# u(A, v, sym = TRUE, ysym = FALSE)
# u(A, v, sym = TRUE, tex = TRUE)

alpha <- 0
beta <- 1
mux <- 0.50
u(Ryacas::ysym(A), v)
u(Ryacas::ysym(A), v, tex = TRUE)
u(Ryacas::ysym(A), v, ysym = FALSE)
u(Ryacas::ysym(A), v, str = FALSE)
eval(u(Ryacas::ysym(A), v, str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2),
234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
