% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\alias{u.default}
\alias{u.yac_symbol}
\title{Vector of Mean Structure Parameters \eqn{\mathbf{u}}}
\usage{
u(A, v, ...)

\method{u}{default}(A, v, ...)

\method{u}{yac_symbol}(A, v, exe = TRUE, R = FALSE, format = "ysym", simplify = FALSE, ...)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{v}{vector of length \code{t} or \verb{t by 1} matrix.
Expected values.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE},
executes the resulting \code{yacas} expression.
If \code{exe = FALSE},
returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE},
the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{R}{Logical.
If \code{R = TRUE}, returns symbolic result as an \code{R} expression.
If \code{R = FALSE}, returns symbolic result as \code{"ysym"}, \code{"str"},  or \code{"tex"}
depending of \code{format}.}

\item{format}{Character string.
Only used when \code{R = FALSE}.
If \code{format = "ysym"},
returns symbolic result as \code{yac_symbol}.
If \code{format = "str"},
returns symbolic result as a characetr string.
If \code{format = "tex"},
returns symbolic result as LaTeX math.}

\item{simplify}{Logical. Simplify symbolic results.}
}
\value{
\eqn{\mathbf{u} = \mathbf{E}^{-1} \mathbf{v}}
}
\description{
Derives the vector of mean structure parameters \eqn{\mathbf{u}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of mean structure parameters \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{u}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)
    \mathbf{v} \\
    =
    \mathbf{E}^{-1}
    \mathbf{v}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{v}_{t \times 1}} vector of expected values.
}
}
\examples{
# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
v <- c(0.50, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
u(A, v)
# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
v <- c("alpha+beta*mux", "mux", 0)
u(Ryacas::ysym(A), v)
u(Ryacas::ysym(A), v, format = "tex")
u(Ryacas::ysym(A), v, format = "ysym")
u(Ryacas::ysym(A), v, R = TRUE)

# Assigning values to symbols

alpha <- 0
beta <- 1
mux <- 0.50
u(Ryacas::ysym(A), v)
u(Ryacas::ysym(A), v, format = "tex")
u(Ryacas::ysym(A), v, format = "ysym")
u(Ryacas::ysym(A), v, R = TRUE)
eval(u(Ryacas::ysym(A), v, R = TRUE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{v}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
