% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\alias{v.default}
\alias{v.yac_symbol}
\title{Vector of Expected Values \eqn{\mathbf{v}}}
\usage{
v(A, u, ...)

\method{v}{default}(A, u, ...)

\method{v}{yac_symbol}(
  A,
  u,
  exe = TRUE,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)
}
\arguments{
\item{A}{\verb{t by t} matrix \eqn{\mathbf{A}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{u}{vector of length \code{t} or \verb{t by 1} matrix.
Mean structure parameters.}

\item{...}{...}

\item{exe}{Logical.
If \code{exe = TRUE}, executes the resulting \code{yacas} expression.
If \code{exe = FALSE}, returns the resulting \code{yacas} expression as a character string.
If \code{exe = FALSE}, the arguments \code{str}, \code{ysym}, \code{simplify}, and \code{tex}, are ignored.}

\item{str}{Logical.
If \code{str = TRUE}, returns symbolic result as string.
If \code{str = FALSE}, returns symbolic result as an \code{R} expression.}

\item{ysym}{Logical.
If \code{ysym = TRUE}, returns symbolic result as \code{yac_symbol} if \code{str = TRUE}.
If \code{ysym = FALSE}, returns symbolic result as string if \code{str = TRUE}.}

\item{simplify}{Logical. Simplify symbolic results.}

\item{tex}{Logical.
If \code{tex = TRUE} returns results as latex math if \code{str = TRUE}.
If \code{tex = FALSE}, returns symbolic result as string if \code{str = TRUE}.}
}
\value{
\eqn{\mathbf{v} = \mathbf{E} \mathbf{u}}
}
\description{
Derives the vector of expected values \eqn{\mathbf{v}}
using the Reticular Action Model (RAM) notation.
}
\details{
The vector of expected values \eqn{\mathbf{v}}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \mathbf{v}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{\mathsf{T}}
    \mathbf{u} \\
    =
    \mathbf{E}
    \mathbf{u}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{t \times t}} represents asymmetric paths
(single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{t \times t}} represents an identity matrix, and
\item \eqn{\mathbf{u}_{t \times 1}} vector of parameters
for the mean structure.
}
}
\examples{
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
v(A, u)

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
u <- c("alpha", "mux", 0)
v(Ryacas::ysym(A), u)
v(Ryacas::ysym(A), u, tex = TRUE)
v(Ryacas::ysym(A), u, ysym = FALSE)
v(Ryacas::ysym(A), u, str = FALSE)

alpha <- 0
beta <- 1
mux <- 0.50
v(Ryacas::ysym(A), u)
v(Ryacas::ysym(A), u, tex = TRUE)
v(Ryacas::ysym(A), u, ysym = FALSE)
v(Ryacas::ysym(A), u, str = FALSE)
eval(v(Ryacas::ysym(A), u, str = FALSE))
}
\references{
McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties
of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology},
\emph{37} (2), 234--251.
\url{https://doi.org/10.1111/j.2044-8317.1984.tb00802.x}
}
\seealso{
Other RAM matrices functions: 
\code{\link{C}()},
\code{\link{Expectations}()},
\code{\link{E}()},
\code{\link{IminusA}()},
\code{\link{M}()},
\code{\link{S}()},
\code{\link{g}()},
\code{\link{u}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{RAM matrices functions}
\keyword{ram}
